+++ packages/ketcher-react/src/script/ui/component/view/saveButton.types.ts
@@ -1,0 +1,5 @@
+export type FileSaverReturnType = Promise<
+  (data: Blob | string, fn, type: string | undefined) => void | never
+>
+
+export type SaverType = Awaited<FileSaverReturnType>
