+++ packages/ketcher-react/src/script/ui/views/components/ContextMenu/ContextMenuTrigger.tsx
@@ -20,14 +20,14 @@ import { useContextMenu } from 'react-contexify'
 import { useAppContext } from 'src/hooks'
 import Editor from 'src/script/editor'
 import {
-  ClosestItem,
-  CONTEXT_MENU_ID,
   ContextMenuShowProps,
-  ContextMenuTriggerType,
-  GetIsItemInSelectionArgs
+  ContextMenuTriggerType
 } from './contextMenu.types'
-import { Selection } from '../../../../editor/Editor'
-import { onlyHasProperty } from './utils'
+import {
+  getMenuPropsForClosestItem,
+  getIsItemInSelection,
+  getMenuPropsForSelection
+} from './ContextMenuTrigger.utils'
 
 const ContextMenuTrigger: React.FC<PropsWithChildren> = ({ children }) => {
   const { getKetcherInstance } = useAppContext()
@@ -141,145 +141,6 @@ const ContextMenuTrigger: React.FC<PropsWithChildren> = ({ children }) => {
       {children}
     </div>
   )
-}
-
-const getIsItemInSelection = ({
-  item,
-  selection,
-  selectedSGroupsIds,
-  selectedFunctionalGroups
-}: GetIsItemInSelectionArgs): boolean => {
-  if (!item || !selection) {
-    return false
-  }
-
-  switch (item.map) {
-    case 'sgroups':
-      return selectedSGroupsIds.has(item.id)
-
-    case 'functionalGroups':
-      return Array.from(selectedFunctionalGroups.keys()).includes(item.id)
-
-    default:
-      return (
-        item.map in selection &&
-        Array.isArray(selection[item.map]) &&
-        selection[item.map].includes(item.id)
-      )
-  }
-}
-
-function getMenuPropsForClosestItem(
-  editor: Editor,
-  closestItem: ClosestItem
-): ContextMenuShowProps | null {
-  const struct = editor.struct()
-
-  switch (closestItem.map) {
-    case 'bonds': {
-      const functionalGroup = FunctionalGroup.findFunctionalGroupByBond(
-        struct,
-        struct.functionalGroups,
-        closestItem.id,
-        true
-      )
-
-      return functionalGroup === null
-        ? {
-            id: CONTEXT_MENU_ID.FOR_BONDS,
-            bondIds: [closestItem.id]
-          }
-        : {
-            id: CONTEXT_MENU_ID.FOR_FUNCTIONAL_GROUPS,
-            functionalGroups: [functionalGroup]
-          }
-    }
-
-    case 'atoms': {
-      const functionalGroup = FunctionalGroup.findFunctionalGroupByAtom(
-        struct.functionalGroups,
-        closestItem.id,
-        true
-      )
-
-      return functionalGroup === null
-        ? {
-            id: CONTEXT_MENU_ID.FOR_ATOMS,
-            atomIds: [closestItem.id]
-          }
-        : {
-            id: CONTEXT_MENU_ID.FOR_FUNCTIONAL_GROUPS,
-            functionalGroups: [functionalGroup]
-          }
-    }
-
-    case 'sgroups':
-    case 'functionalGroups': {
-      const sGroup = struct.sgroups.get(closestItem.id)
-      const functionalGroup = FunctionalGroup.findFunctionalGroupBySGroup(
-        struct.functionalGroups,
-        sGroup
-      )
-
-      return functionalGroup
-        ? {
-            id: CONTEXT_MENU_ID.FOR_FUNCTIONAL_GROUPS,
-            functionalGroups: [functionalGroup]
-          }
-        : null
-    }
-
-    default:
-      return null
-  }
-}
-
-const IGNORED_MAPS_LIST = ['enhancedFlags']
-
-function getMenuPropsForSelection(
-  selection: Selection | null,
-  selectedFunctionalGroups: Map<number, FunctionalGroup>
-): ContextMenuShowProps | null {
-  if (!selection) {
-    return null
-  }
-
-  const bondsInSelection = 'bonds' in selection
-  const atomsInSelection = 'atoms' in selection
-
-  if (selectedFunctionalGroups.size > 0) {
-    const functionalGroups = Array.from(selectedFunctionalGroups.values())
-    return {
-      id: CONTEXT_MENU_ID.FOR_FUNCTIONAL_GROUPS,
-      functionalGroups
-    }
-  } else if (bondsInSelection && !atomsInSelection) {
-    return {
-      id: CONTEXT_MENU_ID.FOR_BONDS,
-      bondIds: selection.bonds,
-      extraItemsSelected: !onlyHasProperty(
-        selection,
-        'bonds',
-        IGNORED_MAPS_LIST
-      )
-    }
-  } else if (atomsInSelection && !bondsInSelection) {
-    return {
-      id: CONTEXT_MENU_ID.FOR_ATOMS,
-      atomIds: selection.atoms,
-      extraItemsSelected: !onlyHasProperty(
-        selection,
-        'atoms',
-        IGNORED_MAPS_LIST
-      )
-    }
-  } else {
-    return {
-      id: CONTEXT_MENU_ID.FOR_SELECTION,
-      bondIds: selection.bonds,
-      atomIds: selection.atoms
-    }
-  }
 }
 
 export default ContextMenuTrigger
